requirements.txt
cachetools==2.1.0
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
httplib2==0.11.3
kafka-python==1.4.3
oauth2client==4.1.2
pyasn1==0.4.3
pyasn1-modules==0.2.2
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
